Minor fixes to typing, move AliasTest class
